Unified forms into one file
