+++ src/app/page.tsx
@@ -1,17 +1,9 @@
-import JobPostingForm from '../components/molecules/JobPostingForm/index'
-import JobPostingForm2 from '../components/molecules/JobPostingForm2/resumo'
-import JobPostingForm3 from '@/components/molecules/JobPostingForm3/empresa'
-import JobPostingForm4 from '@/components/molecules/JobPostingForm4/resumo2'
-
+import JobForm from '../components/molecules/JobForm/index'
 
 export default function Home() {
   return (
     <>
-    <JobPostingForm2></JobPostingForm2>
-    <JobPostingForm3></JobPostingForm3>
-    <JobPostingForm></JobPostingForm> 
-    <JobPostingForm4></JobPostingForm4>
+    <JobForm></JobForm> 
     </>
-    
   )
 }
